requirements.txt
livekit-agents>=1.0.0
livekit-plugins-google>=0.3.0
livekit-plugins-silero>=0.3.0
livekit-plugins-openai>=0.3.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
aiohttp>=3.9.3
aiohttp-cors>=0.7.0
email-validator>=2.1.0
